+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased](https://github.com/poveden/EliteChroma/compare/v1.0.1...HEAD)
+
+### Added
+
+- Add `ChromaAppInfo.xml`
+
 ## [1.0.1](https://github.com/poveden/EliteChroma/compare/v1.0.0...v1.0.1) — 2019-12-29
 
 ### Fixed