+++ CHANGELOG.md
@@ -16,6 +16,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - Add `ChromaAppInfo.xml`
 - Add help links to game folder settings dialog box
 
+### Changed
+
+- Apply better workaround for Chroma startup wait
+
 ## [1.0.1](https://github.com/poveden/EliteChroma/compare/v1.0.0...v1.0.1) — 2019-12-29
 
 ### Fixed