+++ CHANGELOG.md
@@ -9,11 +9,17 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 - Settings dialog box now validates correctly the selected game options folder
+- Fix crash while switching bindings in game
 
 ### Added
 
 - Add `ChromaAppInfo.xml`
 - Add help links to game folder settings dialog box
+
+### Changed
+
+- Apply better workaround for Chroma startup wait
+- Effects are now applied only when the game is in the foreground
 
 ## [1.0.1](https://github.com/poveden/EliteChroma/compare/v1.0.0...v1.0.1) — 2019-12-29