+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased](https://github.com/poveden/EliteChroma/compare/v1.1.0...HEAD)
+
+### Fixed
+
+- Fix silent crash when the game install folder cannot be found
+- Fix main process not exiting when `ValidateFolders()` returns `false`
+
 ## [1.1.0](https://github.com/poveden/EliteChroma/compare/v1.0.1...v1.1.0) — 2020-01-13
 
 ### Fixed